Rename the `src/core/obj.js` file to `src/core/catalog.js`
Now that only the `Catalog` remains in this file, after the previous patches, it makes sense to rename the file to reduce confusion.
